create http distance trait
